fix: allow Integer for progress param in creative_create_service (#701)
* chore: bump rails_mcp_engine to 0.4.1 (fixes float type mapping)

* fix: allow Integer for progress param in creative_create_service

Change Sorbet type from T.nilable(Float) to T.nilable(Numeric) to accept both Integer and Float values.

+++ engines/collavre/app/services/collavre/tools/creative_create_service.rb
@@ -15,7 +15,7 @@ module Tools
     tool_param :after_id, description: "ID of a sibling Creative to insert after. Used for ordering.", required: false
     tool_param :before_id, description: "ID of a sibling Creative to insert before. Used for ordering.", required: false
 
-    sig { params(description: String, parent_id: T.nilable(Integer), progress: T.nilable(Float), after_id: T.nilable(Integer), before_id: T.nilable(Integer)).returns(T::Hash[Symbol, T.untyped]) }
+    sig { params(description: String, parent_id: T.nilable(Integer), progress: T.nilable(Numeric), after_id: T.nilable(Integer), before_id: T.nilable(Integer)).returns(T::Hash[Symbol, T.untyped]) }
     def call(description:, parent_id: nil, progress: nil, after_id: nil, before_id: nil)
       raise "Current.user is required" unless Current.user
